+++ src/powercrud/templatetags/powercrud_partials.py
@@ -1,0 +1,10 @@
+"""Compatibility template tags for Django template partials."""
+
+import django
+from django import template
+
+
+if django.VERSION < (6, 0):
+    from template_partials.templatetags.partials import register
+else:
+    register = template.Library()